Moved back pass_config.hpp; Added doxygen documentation for new class and methods
